+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Uncommented and enabled (by default) KRaft-related metrics in the `kafka-metrics.yaml` example file.
 * Added support for configuring the quotas plugin with type `strimzi` or `kafka` in the `Kafka` custom resource.
 * The Strimzi Quotas plugin version was updated to 0.3.1.
+* Added support to `oauth` authentication for autoconfiguring the use of Kubernetes API server as an authorization server, and the use of service account tokens for authentication.
 
 ### Changes, deprecations and removals